Pre-flood recovery

+++ heuristic_footstep_planner.hpp
@@ -2,4 +2,4 @@
 #include <Eigen/Dense>
 using namespace Eigen;
 
-Eigen::Vector3f desired_footstep_position(const Vector3f& p_ref, const Vector3f& v_CoM, const float& foot_index);
+Eigen::Vector3f desired_footstep_position(const Vector3f& p_ref, const Vector3f& v_CoM, const double& foot_index);
